progress

+++ db/migrate/20201130232259_create_reviews.rb
@@ -4,7 +4,7 @@ class CreateReviews < ActiveRecord::Migration[6.0]
       t.references :user, null: false, foreign_key: true
       t.references :car, null: false, foreign_key: true
       t.text :content
-      t.integer :rating
+      t.float :rating
      
     end
   end
